Setup for Block and Multiblock sounds
This will be used in the future for all machines to have a "i am doing things"-sound.

Also, TwistedGate#214 suggestion happened while I was in the middle of working on this so that's quite the coincidence.
So I have added some things that should aid in doing that as well, sort of.